Initial commit of QC fix
This is an attempt at trying to fix the broken and old QC code in the
repository.

Currently this is a restructuring of the code and an update to use a
newer version fo QC.

Some of the old code was kind of stupid and coun't generate correct
testcases and made "timeouts". Thus we fixed this part and the test are
now running at a reasonable speed. However in the process we have found
some cases that fails. These cases are however not that common and will
only show when running the QC code for a few times.

It is currently uncertain where the flaw lies. It may very well be in
the updated QC code. Thus this needs to be investigated further.

Output from QC has been added to the ERRORS file, where the stackset can
be inspected.